+++ pulse-binding/src/sample.rs
@@ -269,13 +269,6 @@ impl Spec {
         unsafe { capi::pa_sample_spec_valid(self.as_ref()) != 0 }
     }
 
-    /// Checks if the two sample type specifications match.
-    #[inline(always)]
-    #[deprecated(since = "2.7.0", note="use the `PartialEq` implementation instead")]
-    pub fn equal_to(&self, to: &Self) -> bool {
-        self.eq(to)
-    }
-
     /// Gets the amount of bytes that constitute playback of one second of audio, with the specified
     /// sample type.
     #[inline]