+++ pulse-binding/src/channelmap.rs
@@ -41,8 +41,6 @@
 //! [`Map::init_auto`]: struct.Map.html#method.init_auto
 //! [`Map::init_extend`]: struct.Map.html#method.init_extend
 
-#![allow(deprecated)]
-
 use std::borrow::{Borrow, BorrowMut};
 use std::ffi::{CStr, CString};
 use std::borrow::Cow;
@@ -188,13 +186,9 @@ impl From<capi::pa_channel_position_t> for Position {
 pub struct Map {
     /* NOTE: This struct must be directly usable by the C API, thus same attributes/layout/etc */
     /// Number of channels mapped.
-    //TODO: make non-pub (backwards compatible break)
-    #[deprecated(since = "2.10.0", note="do not access directly, use methods from now on")]
-    pub channels: u8,
+    channels: u8,
     /// Channel labels.
-    //TODO: make non-pub (backwards compatible break)
-    #[deprecated(since = "2.10.0", note="do not access directly, use methods from now on")]
-    pub map: [Position; sample::CHANNELS_MAX],
+    map: [Position; sample::CHANNELS_MAX],
 }
 
 impl Borrow<[Position]> for Map {