+++ pulse-binding/CHANGELOG.md
@@ -4,6 +4,7 @@
    (Direct access had been deprecated previously).
  * Removed deprecated comparison methods
  * Removed deprecated proplist methods
+ * Removed deprecated Cargo features
 
 # 2.15.1 (April 18th, 2020)